% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_table.R
\name{create_stanard_kpi_table}
\alias{create_stanard_kpi_table}
\title{Create a KPI table with DAU and DAC averaged over weeks}
\usage{
create_stanard_kpi_table(con, dates)
}
\arguments{
\item{con}{A connection to the \strong{intendo} database. We can use either of the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions to generate this.}

\item{dates}{A vector of dates.}
}
\description{
The \code{create_stanard_kpi_table()} function will create a standardized \strong{gt}
table with DAU and DAC averaged over standard week numbers (where weeks begin
on Sunday). A sequence of \code{dates} should be supplied over a span of time.
For inclusion in reports, the dates should usually go back 60 days from the
present day. Incomplete weeks will be culled from the table.
}
\examples{
\dontrun{
library(lubridate)

# Create a DB connection
intendo <- db_con()

# Get a vector of dates over 30 days
end_date <- as.Date("2015-03-01")

dates <-
  seq(
     from = end_date - lubridate::days(30),
     to = end_date,
     by = 1
     ) \%>\%
  as.character()

# Get the KPI table for a 30-day period
kpi_table <-
  create_stanard_kpi_table(
    con = intendo,
    dates = dates
  )
}

}
