% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{get_dau}
\alias{get_dau}
\title{Get the average DAU value over a particular set of dates}
\usage{
get_dau(con, dates, segment = NULL)
}
\arguments{
\item{con}{A connection to the \strong{intendo} database. We can use either of the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions to generate this.}

\item{dates}{A vector of dates.}

\item{segment}{An optional call to the \code{\link[=segment_daily_users]{segment_daily_users()}} function,
which will segment the \code{daily_users} DB table in order to return a
segmented DAU value.}
}
\description{
This function uses data from the \strong{intendo} database to get the DAU (daily
active users) on a single day, or, the average over a set of days. You need
to supply the database connection object and a vector of dates in ISO format.
}
\details{
We can choose to get a segmented value for DAU (by default, no segmentation
occurs). Support is currently limited, see the help file for the
\code{\link[=segment_daily_users]{segment_daily_users()}} function for details on which segmentation options
are available.
}
\examples{
# Create a DB connection
intendo <- db_con()

# Get the DAU over a three-day period
get_dau(
  con = intendo,
  dates = c("2015-01-01", "2015-01-02", "2015-01-03")
)

}
