% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intendo-package.R
\docType{package}
\name{intendo-package}
\alias{intendo}
\alias{intendo-package}
\title{intendo: A package for Intendo}
\description{
Access to our company data. We can easily authenticate and directly
    use our 'intendo' database tables with the 'dplyr' package. KPIs such as
    DAU, MAU, DAC, and ARPU can be calculated for a collection of dates.
}
\author{
\strong{Maintainer}: First Last \email{email@example.com}

Other contributors:
\itemize{
  \item Intendo [copyright holder, funder]
}

}
\keyword{internal}
