% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{segment_daily_users}
\alias{segment_daily_users}
\title{Segment the \code{daily_users} table in one or more ways}
\usage{
segment_daily_users(con, is_customer = NULL, high_ad_rev = NULL)
}
\arguments{
\item{con}{A connection to the \strong{intendo} database. We can use either of the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions to generate this.}

\item{is_customer}{Segment by whether or not (\code{TRUE}/\code{FALSE}) users are
customers.}

\item{high_ad_rev}{Segment by whether or not (\code{TRUE}/\code{FALSE}) users have
accrued more than $2.00 of ad revenue.}
}
\description{
This function works in tandem with the \code{segment} argument (which may be
available in certain KPI-reporting functions). So far, we can segment with
the following options:
\itemize{
\item is_customer: whether the user is a customer (i.e., ever made a purchase
through an in-app purchase).
\item high_ad_rev: whether a user has earned Intendo high ad revenue (above
$2.00).
}
}
\examples{
# Create a DB connection
intendo <- db_con()

# Get the average DAU value for a three-day period;
# DAU is segmented to those users that gave high
# ad revenue
get_dau(
  con = intendo,
  dates = c("2015-08-01", "2015-08-02", "2015-08-03"),
  segment = segment_daily_users(con = intendo, high_ad_rev = TRUE)
)

}
